requirements.txt
python-dotenv~=0.13.0
datetime~=4.3
psycopg2-binary~=2.8.5
